backend/app/core/security.py
"""JWT 토큰 생성/검증 및 비밀번호 Bcrypt 해싱."""

from datetime import datetime, timedelta, timezone

import bcrypt
from jose import JWTError, jwt

import os
import secrets

# 환경변수 JWT_SECRET_KEY가 있으면 고정 키 사용 (프로덕션),
# 없으면 서버 시작마다 랜덤 생성 → 재시작 시 기존 세션 자동 무효화 (개발)
SECRET_KEY = os.getenv("JWT_SECRET_KEY") or secrets.token_urlsafe(32)
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 60


def hash_password(password: str) -> str:
    return bcrypt.hashpw(password.encode(), bcrypt.gensalt()).decode()


def verify_password(plain: str, hashed: str) -> bool:
    return bcrypt.checkpw(plain.encode(), hashed.encode())


def create_access_token(data: dict, expires_delta: timedelta | None = None) -> str:
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_access_token(token: str) -> dict | None:
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except JWTError:
        return None
