backend/pyproject.toml
[project]
name = "farmos-backend"
version = "0.1.0"
description = "FarmOS 2.0 Backend API — IoT sensor data, irrigation control, and farm management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
